% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgprop.R
\name{mgprop}
\alias{mgprop}
\title{Group Studies with Single Proportions by Homogeneity}
\usage{
mgprop(data, event, n, studlab, ...)
}
\arguments{
\item{data}{A data frame containing the meta-analysis data.}

\item{event}{A vector of event counts.}

\item{n}{A vector of corresponding sample sizes.}

\item{studlab}{A vector of study labels.}

\item{...}{Additional arguments passed on to `meta::metaprop`.}
}
\value{
A list containing the final data with subgroup assignments (`data`),
the final `metaprop` model (`model`), and the number of attempts (`attemps`).
}
\description{
This function iteratively assigns studies to subgroups based on a homogeneity
test. The goal is to create statistically homogeneous groups of studies
before performing a final meta-analysis of single proportions.
}
\details{
The algorithm starts with a single study in "group 1". It then processes
each subsequent study, attempting to place it in an existing group. A study
is added to a group only if its inclusion does not result in significant
within-group heterogeneity. If no suitable group is found, a new one is created.
}
\seealso{
\code{\link{meaning}}
}
\author{
Ahmed Abdelmageed \email{ahmedelsaeedmassad@gmail.com}
}
