% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-grouped.R
\name{grouped}
\alias{grouped}
\alias{print.grouped}
\alias{plot.grouped}
\title{Print and Plot Methods for 'grouped' Objects}
\usage{
\method{print}{grouped}(x, ...)

\method{plot}{grouped}(x, ...)
}
\arguments{
\item{x}{An object of class `grouped`.}

\item{...}{Additional arguments (currently unused).}
}
\description{
S3 methods for objects of class `grouped` returned by the `meaning()` function.
\itemize{
  \item `print.grouped` provides a concise summary view of the results.
  \item `plot.grouped` creates a faceted bar chart to visualize the composition of each subgroup.
}
}
\seealso{
\code{\link{meaning}}
}
\author{
Ahmed Abdelmageed \email{ahmedelsaeedmassad@gmail.com}
}
